vercel-python-app/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
python-http-client==3.3.7
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
Werkzeug==2.2.2
zipp==3.10.0
